+++ CHANGELOG.md
@@ -13,8 +13,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Fixed Rokit erroring on first startup due to some directories not yet being created ([#42])
+- Fixed `selene` and other tools not being installable because their releases contain multiple archives / binaries ([#45])
 
 [#42]: https://github.com/rojo-rbx/rokit/pull/42
+[#45]: https://github.com/rojo-rbx/rokit/pull/45
 
 ## `0.1.7` - July 15th, 2024